+++ transports/socket.go
@@ -1,4 +1,4 @@
-package avroipc
+package transports
 
 import (
 	"fmt"
