+++ transports/transport.go
@@ -1,4 +1,4 @@
-package avroipc
+package transports
 
 import (
 	"io"
